+++ src/main/java/com/elmakers/mine/bukkit/magic/Mage.java
@@ -11,6 +11,7 @@ import java.util.List;
 import java.util.Set;
 
 import com.elmakers.mine.bukkit.api.action.GUIAction;
+import com.elmakers.mine.bukkit.api.spell.SpellKey;
 import com.elmakers.mine.bukkit.effect.HoloUtils;
 import com.elmakers.mine.bukkit.effect.Hologram;
 import com.elmakers.mine.bukkit.spell.ActionSpell;
@@ -800,6 +801,14 @@ public class Mage implements CostReducer, com.elmakers.mine.bukkit.api.magic.Mag
             if (newSpell == null || !(newSpell instanceof MageSpell)) return null;
             playerSpell = (MageSpell) newSpell;
             spells.put(newSpell.getKey(), playerSpell);
+
+            SpellKey baseKey = newSpell.getSpellKey();
+            SpellKey upgradeKey = new SpellKey(baseKey.getBaseKey(), baseKey.getLevel() + 1);
+            Spell upgradeSpell = getSpell(upgradeKey.getKey());
+            if (upgradeSpell instanceof MageSpell)
+            {
+                playerSpell.setUpgrade((MageSpell)upgradeSpell);
+            }
         }
         playerSpell.setMage(this);